Resolve some build warnings
Mainly due to changes to CreateGEP and variants - requirement is now to provide
the element type as part of the call (rather than rely on getting this from the
pointer itself).
In a lot of cases the element type is now provided - in others it is derived
from the pointer type first. This latter approach will require modification in
the future.

Also removed a now redundant member variable.