Update SubgroupBroadcast to handle non-valid non-uniform id
In cases where shaders incorrectly supply SubgroupBroadcast op with a
non-uniform id - we can more elegantly handle this by using a waterfall loop.

It is safe to globally apply this change since if id is constant or uniform,
then the backend will remove the waterfall loop.

This is achieved by adding a new SubgrouBroadcastWaterfall variant that's only
used from the OpGroupNonUniformBroadcast translation.