Fix unlinked copy shader.
The copy shader for a Gs output was changing the locations for the
output so they are densly numbered from 0 to n.  This is a problem
because the Fs inputs may not match causing the the Fs to pick up the
wrong value for the input.

To fix this we will do what was done for the Vs output.  The channel for
the output will be the same as the SPIR-V location for the output.