Allow immutable sampler in relocatable shaders
LLPC will currently fall back to full pipeline compilation when it notices an immutable sampler in the pipeline state.  This was because the relocatable shaders try to load the samplers from the descriptor set, but the sampler data was not there.

A recent change to XGL (https://github.com/GPUOpen-Drivers/xgl/pull/129) writes the immutable sampler to the descriptor set.  This means LLPC does not have to fall back to full pipeline compilation anymore.