Always cache the full pipeline
The decision to not cache the full pipeline when using relocatable shader we done without any data.  Looking at it now, we would like to cache the full pipeline because we can still save some time even if all stages and glue code are already cached.

We want to leave an option that can go back to the old behaviour, so we can easily compare the two paths.