Refine the property in GpuProperty for integer dot product
This change is to make the decision of HW supported integer dot product
more scalable. So the original two boolean members -
`GpuProperty::hasIntegerDot` and `GpuProperty::hasMixedIntegerDot` are
replaced by a struct type - `GpuProperty::supportIntegerDotFlag`,
which can include all HW supported properties, such as the component
bitwidth of a vector and both signedness of the two vectors.