Wavefront size can be different to the subgroup size
Usually subgroup size is fixed value during driver initialization, when
subgroup op is used, the wavefront size has to matched subgroup size no
mater the pipeline shader condition. In the case of the subgroup is 64,
and wavefront is 32, we can disable 32 active lanes, use the 32
wavefront to simulate the subgroup 64 ops.