Fix a mistach issue of LLPC/LGC shader stage mask
The method GraphicsShaderCacheChecker::check() is called by the
middle-end pass PatchCheckShaderCache. However, the input stage
mask and the stage hashes are organized according to LGC shader stage
enumerant order. The order is not necessarily identical to that of
LLPC. We have to treat it as LGC stage mask in the function itself.

This issue is latent because we always have the same order. This
assumption is not ture if either LLPC or LGC introduces new stages
while the counterpart is not.

Change-Id: I2176a0fed998f0bc4cb5427f6d96d9b1fe6931f3