Fix synchronization in ShaderCache
This fixes a double lock and a data race in the `ShaderCache` code.

We should use the main cache lock with the condition variable.

Tested with https://github.com/GPUOpen-Drivers/llpc/pull/1489. In addition,
I enabled TSan and added manual assertions in `lock`/`unlockCacheMap`,
and ran an application that compils a few thousands of pipelines.

Once this is in, I'll send a separate PR to reenable the test in #1489.