Fix LGC tests after changes to the AMDGPU_Gfx calling convention.
This patch aims to fix the LGC tests which are failing caused by a
recent LLVM change 76cbe62262a3ccd0604599a1103ce5d38ac0164b.
Here, the AMDGPU_Gfx calling convention was changed so that some
SGPRs are now callee-save, changing the resulting ISA.
The LGC tests are changed in a way so that they work with and
without the LLVM change.