Remove unused output for some unlinked shaders.
The TES inputs and TCS outputs are not matching up when compiling to
unlinked shaders.  This is because the compiler is inconsistent on when
it will try to pack the inputs and outputs.  It is also inconsistent on
how unused TES inputs are mapped.  This change tries to make that this
consistent.

The idea behind this change is to always remove unused output of the
next stage is available, even if we are doing unlinked shaders.