Implement `lgc::enumRange` in terms of `llvm::enum_seq`. NFC.
Now that LLVM supports safe enum iteration with `llvm::enum_seq`: https://reviews.llvm.org/D107378,
use it to implement `lgc::enumRange`. This only simplifies the code and does not change the behavior.

This also fixes a bug with a dangling reference on reverse iteration, which LLPC never ran into.