Simplify the null fragment shader
The current implementation of the null fragment shader is
more involved than it needs to be.  It will load a value
from a dummy input and then export that value.  This all
seems unnecessary.  I propose that we change it to generate
an empty shader.  LowerFragmentColorExport will add a
dummy export if one is needed.

I've also made the code that generates the null fragment
shader static, so it will be easy to call from the elf
linking when needed.