Add per-input entry point parsing for the standalone compiler
This is to support multi-stage shaders in the future.
See `[RFC] amdllpc input handling improvements` on the mailing list for the high-level plan.

Entry points are not yet passed to the compiler, this will be handled in a future PR.
Because of that, I did not update the option description in the program help.