Fix handling of 8 and 16 bit vertex input in fetch shaders
The fetch shaders do not load values that are smaller than 32-bit correctly.  This is because we end up passing the type that is need for the return value, which loses the fact that it is a smaller value, to VertexFetch.  This is fixed by giving the type from the pal metadata, and then casting the result to the vertex fetch.