Use `ResultError` in standalone compiler
This is so that the offline compiler tool (e.g., amdllpc) can decide how
to exactly report errors. Change APIs to return `llvm::Error`, so that we can write
unit test for failure scenarios that would be difficult to excercise in LIT tests.

For the high-level proposal details, see the `[RFC] Improving LLPC Result checking`
thread on the mailing list.

Also:
-  Simplify error reporting in `amdllpc.cpp`
-  Use early exists in `llpcCompilationUtils.cpp`.
-  Add missing `LLPC_NODISCARD` annotations.
-  Make error messages more consistent.
-  Remove redundant error messages.
-  Do not eagerly print so many errors.