Use full pipeline compilation for convert samplers
The data in the descriptorRangeValues for the conver samplers is used
when translating the spir-v to llvm-ir.  This is too big of a change to
be able to handle as a relocatable shader for now.

We can look into handling it by moving that data to LGC and seeing what
we can do about it.