Handle double in vertex fetch shader.
A double in the return type for a function seems to be returned in two spgrs.  This is a problem
a fetch shader that wants return a vertex input that is a double.  This is fixed by make the return
type for the fetch shader contains a vector of two floats instead of a double.