Correctly handle the DescSet relocation
Code was added to generate a relocation for the off of the descriptor
table for a given descriptor set.  However, the code the replace the
relocation was not written.  I cannot remember why, but it was probably
because I could not create a test that actaully needed it.  The code
added and forgotten.  I have not found a case that will generate it.
I'm adding a test and adding the code to replace the relocation.