Fix the parameter type for the merged shaders
When adding the fetch parameter to a merged shader, we simply use the type in
the vertex fetch info.  This can cause a mismatch with the type for the es
(vertex) shader when generating the function call.  To fix this, we need to
use the "vgpr type" that corresponds to the type in the fetch info.