Reorganize shaderdb tests
This divides shaderdb into a few new categories. The goal is to make
working with existing and future tests more manageable and make sure
we don't have hundreds of files in non-leaf project directories.

The new organization is as follows:
```
shaderdb/
 core/                -- Core SPIR-V compilation.
 debug_info/          -- Debug info generation tests.
 error_reporting/     -- Standalone compiler error reporting.
 extensions/          -- SPIR-V/Vulkan extensions.
 fuzzer/              -- Regression tests for issues found by fuzzers.
 general/             -- The remaining uncategorized tests.
 gfx10/               -- GFX10-specific compilation tests.
 gfx9/                -- GFX9-specific compilation tests.
 hlsl/                -- HLSL input tests.
 multiple_inputs/     -- Tests that require multiple input files.
 object/              -- Obj*.pipe tests.
 relocatable_shaders/ -- Relocatable shader ELF compilation and caching tests.
```

Also remove the `testShaders.py` script which is not necessary to run LIT tests.