Don't try reloc compilation with inline buffers
Inline buffers require a bit of special code that the reloc compilation
cannot know will be required until the user data nodes are available.
For now, it will be best if we just avoid reloc compilation when inline
buffers are used.  This allows a large number of VK CTS tests with pass.
We can revisit later if needed.