Add simple parallel for loop implementation
I also considered using `llvm::parallelForEachError`, but run into
errors with the global thread pool strategy. In short, the LLVM
implementation is problematic for a few reasons:
-  It uses a single global thread pool strategy. If we override it
   in LLPC, we are risking messing up with some future parallelism
   in the LLVM code.
-  It does depends on ManagedStatics and the global destruction/shutdown
   order. When used in `amdllpc.cpp` TSan complains about the destructor
   code.
-  The LLVM thread pool implementation seems a somewhat heavyweight as
   it uses multiple synchronization primitives in its implementation.
   I have not benchmarked the implementation from this PR against LLVM,
   but doubt all the complexity makes it faster in my intended simple
   use cases.

Also add unit tests and refactor the ShaderCache test to use the new
implementation.

In the future, I plan to use this to compile multiple amdllpc inputs in
parallel.