Allow parallel pipeline/shader compilation
Parallel compilation is controlled by the `--num-threads` flag.
We default to single-threaded compilation.

The main goal is to excercise the threading in the compiler without
having to run the full ICD on a machine with a GPU. With the thread
sanitizer being enabled in our public CI, we hope that this will catch
most of the potential synchronization issues in LLPC.

Parallel compilation and verbose output are disallowed because of
LLVM's ostream not being thread safe. Even if this wasn't the case,
verbose logs would be interleaved and not very useful.

Refactor stage processing such that `ShaderModuleData` mergin happens
at the end.