Do not use `strtok` in vfxParser
`strtok` relies on global variables and is not thread safe.

This replaces all uses of `strtok` in LLPC with a thread safe helper
function `split`. The semantics are not identical, but the use of `split`
in the parser code preserves the existing behavior.

Add Vfx unit tests.

Also add macros to guards against future accidental uses of `strtok`.