Make assemble spvasm to spv thread safe
Assembling to an output file is now protected with a mutex. This is
required as assembling, outputting and recording the file as assembled
needs to be in a critical section.

Assembling the same file will only happen once.
Opening a file for reading on windows now uses fopen instead of fopen_s
which does not allow shareable (opening more than once). This is in
line with the Linux implementation.