Port more middle-end passes to LLVM's new pass manager
This patch ports more middle-end passes and all the optimization passes to the
new pass manager. Some of the passes use the LegacyDivergenceAnalysis pass.
This LLVM pass is not ported to the new pass manager, so we use the new
DivergenceAnalysis pass when porting these passes to the new pass manager.
I tried comparing the code generated with LegacyDivergenceAnalysis and
DivergenceAnalysis passes and didn't find any difference.