Use in-memory binary for glsl and spvasm intermediate binary spv
Outputting a binary form spv file isn't required by the amdllpc tool. Move to
using an in-memory binary object instead. This should also result in a slight
speed-up as fewer file writes are required (and it is thread safe).

Also update fopen_s to use standard fopen as per Linux implementation to avoid
problems with multiple threads attempting to read the same file on Windows.