Fix relocatable shaders with Tess and Gs
The shader merger does not work correctly when all stages of the graphics pipeline are present.  We should not be passing the vertex inputs as parameters in the merged Tes-Gs shader.  We also need to place the merged Vs-Tcs at the start of the text section so the fetch shader can be correct prepended.