Add pipeline dump recompilation tests
Before this change, we only checked that some pipeline dumps are
produced by amdllpc, but did not attempt any correctness checks.

This patch adds some basic correctness checks. We make sure that
the pipeline state is preserved in the produced `.pipe` files and that
we after feeding them back to amdllpc they produce the same assembly.