Part pipeline: Fix wave size issue in vertex shader
The default wave size for a part pipeline compilation rely on whether
sugroup size is used in the other part. In the whole pipeline mode it
is normally handled by querying the shader modes field, so we need to
make sure that the shader modes have proper information even in the part
pipeline mode.