Part pipeline: Interface changes
This is the first commit in the series to implement LGC part-pipeline
compilation scheme.

The part-pipeline is guarded by the new option EnablePartPipeline, which
is off by default. It is expected that when all the commits in the series
have been pushed, the old partial-pipeline compilation scheme will go
away.

The part-pipeline scheme is as follows.

First it compiles the FS with FS-applicable pipeline state (or finds in
the shader cache). The FS ELF's PAL metadata contains information on FS
input packing for use when compiling the pre-rasterization
part-pipeline.

Then it compiles the pre-rasterization part-pipeline (VS,TCS,TES,GS)
with pre-rasterization pipeline state and the FS input packing
information (or finds it in the shader cache).

Then it uses the LGC ELF linker to link them together.