Separate the collection of info
There is one function that does 3 things.  It verifies that the spirv
binary is well formed.  This is done when validating the spirv just
before, and can be left to that function.

It also computes the total size of all of the debug instructions.  This
is essentially done when the debug code is trimmed, so we will leave it
to that function.  Note that this means the buffer that holds the
trimmed code cannot be allocated to the exact right size, and more
memory will have to be allocated.  I do not believe this to be a problem
because individual shader are usually small, and this buffer is short
lived.  It should not have a noticable effect on performance or memory
consumption.

The last thing it does is to collect the shader module usage info.  This
is function is specialize to do just that one thing.

This should simplify the code.