Rename PassRegistry.inc to PassRegistry.inc (#1753)
Internally, we have bazel build files that build llpc.  The new file PassRegistry.def causes problems because it cannot be included in the "srcs" if a [cc_library rule](https://docs.bazel.build/versions/main/be/c-cpp.html#cc_library.srcs).  If it is not a problem, I would like to change the extension of the file to ".inc".