Set BoundCtrl to true when using DPP for OpGroupNonUniformQuad*
When generating LLVM IR for OpGroupNonUniformQuadBroadcast and
OpGroupNonUniformQuadSwap, set the BoundCtrl argument to the
llvm.amdgcn.mov.dpp intrinsic to true. The effect of this is that if a
write to an active lane depends on the value of an inactive lane, a zero
value will be written, instead of leaving the active lane unchanged. The
SPIR-V spec for these ops says that "if an active invocation reads Value
from an inactive invocation, the resulting value is undefined", so this
should not affect correctness.

This allows the LLVM backend to generate better code, because a
v_mov_b32_dpp instruction with BC=1 can easily be combined into a
following arithmetic instruction by propagating the BC=1 bit to the
combined instruction. With BC=0 it is much harder, and could probably
only be done if the compiler could prove that all active lanes only
read from other active lanes.