Resource Layout scheme
Indirect scheme is designed for the case that the pipeline layout only
contains part of information of the final executable pipeline. So that
user data is used in conservative way to make sure that this pipeline
layout is alaways compatiable with other layouts which may contain the
descriptor which is not known currently.

In indirect mode, the user data registers for various resources are
allocated in the following fashion:
 1. one user data entry for vertex buffer
 2. one user data entry for the push constant buffer
 3. descriptor set index for each set.