Set BoundCtrl to true when using DPP for reductions and scans
When generating LLVM IR for:

OpGroupFMax (and other similar OpGroup* opcodes)
OpGroupNonUniformFMax (and other similar OpGroupNonUniform* opcodes)

set the BoundCtrl argument to the llvm.amdgcn.mov.dpp intrinsic to true,
similar to #1782 which did it for OpGroupNonUniformQuad*.

This makes it easier for the backend to combine the v_mov_b32_dpp into
a following arithmetic instruction. Setting BC=1 does not change the
result because none of the DPP operations read from an inactive lane
because reductions execute in WWM. LLPC knows this, but that knowledge
is not so readily available in the LLVM backend.

Inclusive and exclusive scans use some DPP operations like row_shr:1
which read from invalid lanes, so we can't set BC=1 on them, but they
also use operations like row_bcast:15 where we can set BC=1.