Add a vkgc_headers library to make defines consistant
Currently, defines like LLPC_CLIENT_INTERFACE_MAJOR_VERSION need to be
defined on every library that includes vkgcDefs.h. If it is not defined
to the same value on every linked library, hard to debug bugs will
appear because defined structs end up with different sizes.

To fix that, add an interface library in CMake called vkgc_headers that
only exposes the header file and the needed defines to get consistent
definitions.

Use this library instead of adding the include path directly.
Also add some temporary workaround to decouple the llpc change from the
spvgen change.