SPIRVReader: Fix a padding issue of OpRuntimeArray
When translating OpRuntimeArray in spir-v reader, there is an assumption
for required padding cases that the array stride should be greater than
the size of the array element. Howerver, there is an exception in HLSL
that the array element is a structure with variable-sized array as the
last member, which should not be identified as a padding case as the
behavior of GLSL.
To fix that, the assertion is removed and a new condition is added to
block this kind of arrays to be padded.
Add the corner case as a lit test in shaderbd\hlsl.

Fixes: assertion on dEQP-VK.tessellation.fractional_spacing.glsl_*