lgc/patch: Integrate PatchBufferOp into the post-NGG part of the pass manager
There is no good reason to run a partial InstCombine just after
PatchBufferOp if we're going to run InstCombine after NGG-merge
anyway.

Also, PatchBufferOp can easily be moved later in the pipeline,
which moves it out of the way of the more complex module-level
passes. That will hopefully make it easier to clean up in that
area.

Part-of: https://github.com/GPUOpen-Drivers/llpc/issues/1892