lgc/patch: Add more `const` in PatchInOutImportExport
This fell out of a review of this pass's place in the compilation
pipeline.

Unify some relatively redundant code paths. ResourceUsage::inOutUsage.gs.builtInOutLocs
becomes a std::map for consistency. This actually fixes a potential issue:
that map is fed into the cache hash calculations, so it's better
to use a map with guaranteed deterministic ordering.

v2:
- fix a coding style issue