lgc/patch: Move IPSCCP pass earlier
There seems to be no particularly strong reason for its current placement.
The main motivation of the move is to move it out of the way for future
refactoring.

To be honest, it's kind of unclear to me why we're using this pass in
the first place. Is there any example where we compile anything that
could benefit from generic inter-procedural constant propagation?

v2:
- test fix: ensure that lgc.get.desc.ptr doesn't become undef