lgc/patch: Rearrange the pass pipeline
This major restructuring of the pass pipeline follows the principle
that passes should be ordered as follows:

1. Optimizations running on "higher-level" IR, meaning in particular
   that GPU concepts are visible at a high level. This constitutes

 - Generic LLVM canonicalization (in particular InstCombine) and
   optimizations (SimplifyCFG, CSE, loop optimizations, ...)
 - GPU-specific optimization passes, e.g. cross-shader optimizations,
   the (rather unfortunately named) PatchReadFirstLane pass, etc.

2. Lowering of the IR into "AMDGCN-style" IR that is suitable for
   backend consumption, meaning:

 - There are no lgc.* intrinsics anymore
 - llvm.amdgcn.* instrinsics are used instead
 - Functions have the correct AMDGPU calling convention and signatures

3. "Cleanup" optimizations that remove inefficiencies that were
   introduced by the lowering. This should *not* be a full
   optimizations run but should be specifically aimed at issues
   created by lowering.

This change by itself doesn't quite fulfill the promise yet because
a lot of lowering actually happens in BuilderReplayer which this
change leaves at the top (for good reason: "recorded" calls are quite
awkward to work with). Improving this situation is left for future
changes, ideally building on the proposed llvm-dialects library.

Part-of: https://github.com/GPUOpen-Drivers/llpc/issues/1892