Add support for scalarize waterfall descriptor loads.
Provide support for scalarizing waterfalled non-uniform descriptor
loads in limited circumstances.
This is an alternative to the current behaviour which scalarizes
access to a vector loaded outside the waterfall loop.

While using scalarized loads introduces memory access into each
loop iteration it significantly reduces VGPR pressure created by
the loop.
It also has the positive side effect of reducing the probability
that a s_waitcnt on the vmcnt is introduced into the waterfall loop
by the backend, as lgkmcnt is used for the loads instead.

In practice performance of the scalarized loads can be equivalent
to vector loads when there is medium to high commonality in the
non-uniform values.

Automatically enable this scalarization when a vgpr limit is set.
As this is where it will be most beneficial.

The limitation of this transform is that loads are only scalarized
when all non-uniform values have a common base/index.
This covers the majority of scenarios.
For a general purpose solution instruction would need to be heavily
reordered to ensure the correct parts occur inside/outside the
waterfall loop.