Add disableFastMathFlags mask shader option (#1818)
Add a new shader option that can be used to disable fastMathFlags when
translating from spirv to llvm-ir.

In addition, the option is encoded in the .pipe file (as well as the previously
created fastMathFlags option).
The new option is preferred as it allows selective disabling of the fast math
flags after the fastMathFlags are calculated in the normal way. This is more
selective than the previous fastMathFlags option.