Add mesh shader support: Rename vertexIndex to vertexOrPrimitiveIndex
In the builder methods 'CreateWriteGenericOutput' and
'CreateWriteBuiltInOutput', they have a parameter 'vertexIndex' for TCS
output to represent the outermost dimension as vertex indexing. Now, the
two methods will be used by mesh shader. And mesh shader introduces
per-primitive outputs which need this parameter and the semantics are
similar. Rename it to 'vertexOrPrimitiveIndex' to accommodate this
change.