llpc/SPIRVReader: Ensure allocas are at the start of the entry block
As explained in the comment, we want to ensure that (entry block) allocas
always stay in the entry block, even if a later transform (e.g.
inlining) causes the entry block to be split.

Note that this only applies to the `alloca` instruction itself, not to
the store that initializes it.

It also wouldn't apply to "true allocas" that appear in other basic
blocks, though we should never create those anyway since SPIR-V doesn't have an
equivalent of the C alloca() function.