lgc: add BuilderCommon::setInsertPointPastAllocas
This method places the builder's insert point between entry point
allocas and other instructions. That's the point where one should insert
instructions that define values that should be visible to the entire
function. This avoids splitting up the initial block of allocas, which
in turn avoids allocas moving out of the entry block e.g. by a later
inlining transform.