lgc/PassManager: use report_fatal_error for bad command-line options
llvm_unreachable() is like assert() in that it should only be used to
flag conditions that are programmer error. Depending on the build
options, the compiler may actually optimize code based on the assumption
that it is never reached, and so paths that would reach it become
undefined behavior.

The arguably correct thing to do here is a report_fatal_error(), because
the condition can only be reached via bad debug settings.