Do not expand gep change - remove instruction in correct order
PR #1833 implemented a significant cleanup of llpcSpirvLowerMemoryOp. However,
part of that change removed the 2 start instruction removal that's required to
stop instructions with uses from being removed, resulting in asserts.

This is simply an ordering issue since pairs of instructions are added to the
removal list, and the ordering is important.

Re-instate the 2 stage removal process from the original code.