[CI] Add script to garbage collect old amdvlk docker images (#1881)
The script works for the Google Cloud Artifact Registry.

I have also considered using the gcloud python API instead of invoking `gcloud` as a subprocess, but that proved much more difficult and verbose. Invoking `gcloud` also has the advantage that it makes it easy to manually replay the commands when necessary.

As a next step, we should make the nightly CI workflow run the script.