[CI] Automatically garbage collect old docker images (#1885)
Run the gc script in the nightly AMDVLK wokflow, but only in the
'+clang' configuration. This is so that we do not have multiple workers
trying to delete the same images. We also expect the '+clang'
configuration to be the most robust so that it gets to the end of the
workflow consistently.

Update the gc script to be compatible with python 3.8, which is
installed by default on ubuntu 20.04.