Stop relying on s_barrier implying s_waitcnt
LLVM may not further generate s_waitcnt for a s_barrier, LLPC need to
insert explicit fence wherever it needs.

Per Nicolai's comment:

Technically from a formal memory model perspective, each of those
barriers should have an Acquire fence of the same scope after the
barrier (besides a Release fence before it).

All the fences can be at workgroup scope (instead of system). The reason
is that there is only ever synchronization between threads of the same
workgroup; in some cases because it's anyway LDS, in other cases because
it's synchronization between e.g. LS and HS that are merged together
into the same workgroup.