lgc/ShaderMerger: Builder-ify and factor out some common code
Factor out some common code for handling user data values at the same
time.

This change is meant to not change the generated code. However, it does
end up changing the *attributes* of call-sites of intrinsics. These
attributes are redundant and *shouldn't* be added, but the difference is
enough to cause some minor changes after all. Follow-up changes will
cleanup an instance of a change that I happened to see.

v2:
- rename b --> builder