lgc/PatchEntryPointMutate: don't process calls in graphics shaders
This is (a) unused (except for lit tests) and (b) almost certainly broken,
because the pass as-is does not adjust the signatures of non-entry point
functions in graphics pipelines. Even if we were to see an indirect call
to a function compiled separately, it is exceedingly unlikely that the
argument lists would match.

As such, best to remove this call since it only confuses (and wastes a
little bit of compile time). Once function calls in graphics pipelines
are truly desired, a separate change can do the implementation properly.
Compute pipelines are not affected by this change.