lgc/PatchCheckShaderCache: delete cached functions' bodies
This should really have been done from the very beginning. As-is, the
pass pipeline seems to rely on:

1. IPSCCP cleaning out the function bodies (replacing them by
   unreachable), and
2. a later GlobalDCE pass removing the functions.

That's messy, so let's be cleaner and explicit.

We could probably just delete the functions outright, but let's not do
that for now on the off-chance that the function may have a use
somewhere. Leaving around only a function declaration is cheap enough.

This improves compile-time by avoiding that later passes uselessly
look at the functions' code.