Fix OpAtomic{FMin,FMax} implementation
It turns out that this implementation was always broken because it is
not allowed (and has never been allowed) to use atomicrmw min/max with a
floating point type. This causes an IR verifier error which was
apparently never caught.

LLVM main recently gained the fmin/fmax operations of atomicrmw (in
commit 1023ddaf77901 ("[LLVM] Add the support for fmax and fmin in
atomicrmw instruction"). Their implementation in the AMDGPU backend is
not (yet?) complete, but that's okay: PatchBufferOp does the required
code substitution.

This change fixes the IR verifier issue.

Fixes: dEQP-VK.glsl.atomic_operations.min_float32_vertex