lgc/PatchResourceCollect: don't pack VS-TCS if dynamic indexing is used
Changing the order of passes in the patch pipeline revealed a bug here
in how the VS-TCS interface is packed. Specifically,
updateOutputLocInfoMapWithPack() would fail to remove an output export
instruction, which lead to the same VS-TCS component being written twice
(once with incorrect data).

This was due to how the code tries to conservatively handle the case of
dynamic indexing; but then a lot of the support around dynamic indexing
is probably incorrect in the first place, because the possible range /
the underlying array size is lost when translating from SPIR-V to lgc
intrinsics.

Instead of trying to limp along when dynamic indexing is used, just
don't pack at all in that case -- which allows an easy fix for the bug
that was encountered on the non-dynamically-indexed path.

Fixes: dEQP-VK.spirv_assembly.instruction.graphics.module.vert2_geom2_tessc2_tesse1_frag2