Add SinkingPass after PatchBufferOp (#1902)
The intention is to allow us to stop running the IR SinkingPass as part
of the backend codegen pass pipeline, which is very unusual. Currently
SinkingPass seems to be required to clean up some of the IR generated by
PatchBufferOp so I have placed it there. In future it would be good to
remove it altogether, since it can have bad effects on register
pressure.