lgc/PatchInOutImportExport: fix TCS output locOffset handling
No longer try to special-case constant locOffset values.

The underlying issue was that while `value` was adjusted by `locOffset`,
the ultimate value of `loc` was not adjusted.

It should also be possible to also fix this bug by updating locInfoMapIt
for the new value of `value`; however, I believe it is better to
simplify the code overall by removing the special case.

I encountered this issue while experimenting with alternative pass
orderings.

Fixes: dEQP-VK.tessellation.user_defined_io.per_vertex_block.vertex_io_array_size_spec_min.isolines