Remove the function clearInactiveBuiltInInput in PatchResourceCollect
The ADCE pass in lower has already removed such built-in input dead
calls. There is no need of re-checking them and doing follow-up built-in
usage clean-up work. The codes will not be executed. Ideally, front-end
ought to get rid of all dead calls before the IRs are passed to LGC.