Add option to disable LLVM scalarizer pass.
We found that disabling the scalarizer pass helps
with generating v_pk_* instructions and in some cases
with VGPR usage, but we cannot go with completely
disabling the scalarizer pass for now.
This patch adds a new LLPC CLI option which
can be used to control running the scalarizer
pass, which should help with experiments.