Fix out of bound issue in in reassembleOutputExportCalls (#1977)
Setting the size of `elementsInfoArray` based on the count of output
export calls may cause out of bound issue when the actual used output
attributes in the current stage are not matching the input attributes in
the next stage. To fix it, we should use the size of input atrributes in
the next stage. Besides, modify `InOutPackingNonZeroBase.lgc` to cover
this kind of corner case.