Opaque pointers: Check if GV was created from OpVariable
This function (transShaderDecoration) is called after translation SPIRV
instruction to LLVM instructions and it is updating some of Global
Variable meta data.
Additional check, which ensures that Global Variale was created from
OpVariable needs to be added. This is because for opaque pointer mode
OpAccessChain not always is converted to GetElementPtr. If OpAccessChain
is a zero-index then GetElementPtr instruction will be skipped and
source of OpAccessChain will be taken. From SPIRV to LLVM translation
perspective this will look like OpAccessChain is converted to Global
Variable. Despite the fact that we are skipping here Global Variable
initialization of meta data, this Global Variable was initialized
before, when translation from OpVariable to Global Variable took place.