lgc/dialect: add buffer.length and buffer.ptr.diff
Convert existing "LGC intrinsics" to start a "sub-dialect" for buffer
fat pointers, i.e. LLVM IR pointers whose underlying representation is
a buffer descriptor -- essentially a base + buffer size -- and an
offset into the buffer.

buffer.desc.to.ptr is still missing because of opaque pointer troubles.
Basically, llvm-dialects doesn't know how to handle an operation that
returns a non-opaque pointer, and it doesn't make much sense to spend
effort on adding that support anymore. So I plan to add
buffer.desc.to.ptr once we've fully transitioned to opaque pointers.

Some `ashr / sdiv` become `lshr` in the test cases. This is okay because
the resulting value is truncated.

Update the llvm-dialects submodule to use the new `Memory<...>`
functionality.